BAU: Refactor - move user identity code into useridentity namespace
